[JP-06][API] Renamed filter category file
