+++ src/ts/enums/filter-сategory.ts
